add 'sync/' package to package-wide tsconfig and test runner.
For some reason Orbit had no included these in the normal build + test pipline, by which I
mean that which gets run using 'bun run build' and bun run test'. I added the sync/ package
back in and everything worked. So I'm going to assume this is because Andy forgot this.